Fixed chunksize/number of chunks setting in ORC.
Orc needs exact match between number of chunks (in h2o) and number of stripes.
This can not be achieved just by setting chunk size - we have to set number of chunks explicitly and compute respective chunk size.
Last chunk can be significantly larger, but we do not care about the size in this case, only the number of chunks matters.